+++ R/check.R
@@ -60,7 +60,7 @@ check <- function(bioc = TRUE, timeout = as.numeric(Sys.getenv("R_REVDEPCHECK_TI
   }
 
   envs <- Sys.getenv()
-  envs <- envs[grep("^_?R_CHECK_", names(envs))]
+  envs <- envs[grep("^(_?R_CHECK_.*|NOT_CRAN)$", names(envs))]
   if (length(envs) > 0L) {
     envs <- sprintf(" %02d. %s=%s", seq_along(envs), names(envs), envs)
     envs <- paste(envs, collapse="\n")