+++ R/run.R
@@ -26,9 +26,12 @@
 #' --preinstall-todo     Pre-install all packages to be checked
 #' --preinstall <pkgs>   Pre-install specified packages
 #' --preinstall-children Install all reverse dependencies
+#' --preinstall-grandchildren
+#'                       Install all second-generation reverse dependencies
 #' --preinstall-update   Install packages that have been updated since last run
 #' --preinstall-error    Install packages that gave an "error" during checks
 #' --preinstall-failure  Install packages that failed to be checked
+#' --force               Force re-installation of select packages.
 #'
 #' Add and remove packages to be checked:
 #' --reset               Full reset to restart checks from scratch
@@ -90,6 +93,11 @@ run <- function(pkg = ".", ..., warn = 1L, args = base::commandArgs(trailingOnly
   } else if ("--version" %in% args) {
     cat(as.character(packageVersion(.packageName)), "\n", sep = "")
     return(invisible())
+  }
+
+  force <- FALSE
+  if ("--force" %in% args) {
+    force <- TRUE
   }
 
   assert_repos()
@@ -226,21 +234,25 @@ run <- function(pkg = ".", ..., warn = 1L, args = base::commandArgs(trailingOnly
     if (identical(pkg, ".")) pkg <- revdep_this_package()
     cran_revdeps <- import_from("revdepcheck", "cran_revdeps")
     pkgs <- cran_revdeps(pkg)
-    revdep_preinstall(pkgs)
+    revdep_preinstall(pkgs, skip = !force)
+  } else if ("--preinstall-grandchildren" %in% args) {
+    if (identical(pkg, ".")) pkg <- revdep_this_package()
+    pkgs <- revdep_grandchildren(pkg)
+    revdep_preinstall(pkgs, skip = !force)
   } else if ("--preinstall-error" %in% args) {
     res <- revdepcheck::revdep_summary(pkg)
-    revdep_preinstall(revdep_pkgs_with_status(pkg, "error"))
+    revdep_preinstall(revdep_pkgs_with_status(pkg, "error"), skip = !force)
   } else if ("--preinstall-failure" %in% args) {
     res <- revdepcheck::revdep_summary(pkg)
-    revdep_preinstall(revdep_pkgs_with_status(pkg, "failure"))
+    revdep_preinstall(revdep_pkgs_with_status(pkg, "failure"), skip = !force)
   } else if ("--preinstall-todo" %in% args) {
     pkgs <- todo(print = FALSE)
-    revdep_preinstall(pkgs)
+    revdep_preinstall(pkgs, skip = !force)
   } else if ("--preinstall" %in% args) {
     pos <- which("--preinstall" == args)
     if (pos == length(args)) stop("Missing value for option '--preinstall'")
     pkgs <- parse_pkgs(args[seq(from = pos + 1L, to = length(args))])
-    revdep_preinstall(pkgs)
+    revdep_preinstall(pkgs, skip = !force)
   } else {
     if (length(args) > 0L) {
       stop("Unknown command-line arguments: ",